Brick 15-final: Rebuild database and embeddings
- Fixed 6 descriptions, 2 names, 2 prices via cleanup scripts
- Filled 6 missing restaurant_places_data entries (underscore/hyphen variants)
- Filled 5 missing name_mapping entries
- Rebuilt menu_db.json: 18,376 items from 463 restaurants
- Rebuilt menu_embeddings.npz: 18,376 items via text-embedding-3-large (float16)
- Embeddings stored as float16 to stay under GitHub 100MB limit
- Updated build_vector_db.py to save float16 by default
- All 90 tests passing

+++ backend/scripts/build_vector_db.py
@@ -142,8 +142,8 @@ def main():
     print("Generating embeddings...")
     embeddings = build_embeddings(menu_index)
 
-    # Save embeddings as compressed npz (stays under GitHub 100MB limit)
-    np.savez_compressed(EMBEDDINGS_FILE, embeddings=embeddings)
+    # Save embeddings as compressed npz in float16 (stays under GitHub 100MB limit)
+    np.savez_compressed(EMBEDDINGS_FILE, embeddings=embeddings.astype(np.float16))
     print(f"Saved {EMBEDDINGS_FILE}")
     print(f"Built embeddings for {len(menu_index)} items. Saved to menu_embeddings.npz")
